Separation of user registration and query
